fix: use file name for slug instead of field
