+++ src/lib/posts.ts
@@ -42,15 +42,8 @@ export function fetchPostContent(): PostContent[] {
         fullPath: string;
       };
       matterData.fullPath = fullPath;
-
-      const slug = fileName.replace(/\.mdx$/, "");
-
-      // Validate slug string
-      if (matterData.slug !== slug) {
-        throw new Error(
-          "slug field not match with the path of its content source"
-        );
-      }
+      matterData.slug = fileName.replace(/\.mdx$/, "");
+      console.log(matterData);
 
       return matterData;
     });
